included cloud tag style
Copied the tag could style from theme Berlin and included it in the bigpicture theme which was lacking it initially. The tag cloud group is inserted under the Items/Browse group and included the line-height: 2em so that there is no overlap between the words in adjacent lines. Hope this works!